chore: update tests to run without redis
